api_backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
firebase-admin
openai>=1.0.0
